requirements.txt
pandas==2.1.4
numpy==1.26.2
transformers==4.36.2
sentence-transformers==2.2.2
torch==2.1.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
nltk==3.8.1
tqdm==4.66.1